Plagin ya 2.0

+++ Assets/Scripts/Coin/Wallet/IWallet.cs
@@ -1,6 +1,6 @@
 ﻿public interface IWallet
 {
-    public int CountCoins { get; }
+    public long CountCoins { get; }
 
     public void SetView(IWalletView walletView);
